git-svn-id: https://gorgonlib.googlecode.com/svn/branches/2.x@930 a61f9e97-bd4c-0410-8c65-85b29eaa70ff

+++ Dorian/Gorgon/Gorgon.Graphics/Targets/GorgonSwapChain.cs
@@ -292,7 +292,7 @@ namespace GorgonLibrary.Graphics
 			ReleaseResources();
 			GI.SwapChainFlags flags = GI.SwapChainFlags.AllowModeSwitch;
 
-			GISwapChain.ResizeBuffers(Settings.BufferCount, Settings.VideoMode.Width, Settings.VideoMode.Height, (GI.Format)Settings.VideoMode.Format, (int)flags);
+			GISwapChain.ResizeBuffers(Settings.BufferCount, Settings.VideoMode.Width, Settings.VideoMode.Height, (GI.Format)Settings.VideoMode.Format, flags);
 			CreateResources();
 
 			OnTargetResize();
